+++ pyleco/utils/message_handler.py
@@ -28,7 +28,6 @@ import logging
 import time
 from typing import Any, Callable, Optional, Union
 
-from openrpc import RPCServer
 import zmq
 
 from ..core import COORDINATOR_PORT
@@ -37,6 +36,7 @@ from ..core.message import Message, MessageTypes
 from ..core.serialization import JsonContentTypes, get_json_content_type
 from ..json_utils.errors import JSONRPCError
 from ..json_utils.rpc_generator import RPCGenerator
+from ..json_utils.rpc_server import RPCServer
 from .base_communicator import BaseCommunicator
 from .log_levels import PythonLogLevels
 from .zmq_log_handler import ZmqLogHandler